requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-Migrate>=4.0.5
Werkzeug>=3.0.0
APScheduler>=3.10.0
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0
ldap3>=2.9.1
cryptography>=42.0.0
